+++ src/app.module.ts
@@ -15,7 +15,6 @@ import { RedisModule } from './redis/redis.module';
 import { BoardsModule } from './boards/boards.module';
 import { CacheModule } from '@nestjs/cache-manager';
 import { StorageModule } from './storage/storage.module';
-import { OauthModule } from './oauth/oauth.module';
 import { ElasticsearchModule } from './elasticsearch/elasticsearch.module';
 
 @Module({
@@ -37,7 +36,6 @@ import { ElasticsearchModule } from './elasticsearch/elasticsearch.module';
     RedisModule,
     BoardsModule,
     StorageModule,
-    OauthModule,
     ElasticsearchModule
   ],
   controllers: [AppController],
