--- src/oauth/dto/create-oauth.dto.ts
@@ -1,1 +1,0 @@
-export class CreateOauthDto {}
