--- src/oauth/dto/update-oauth.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/swagger';
-import { CreateOauthDto } from './create-oauth.dto';
-
-export class UpdateOauthDto extends PartialType(CreateOauthDto) {}
