--- src/oauth/oauth.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { OauthController } from './oauth.controller';
-import { OauthService } from './oauth.service';
-
-describe('OauthController', () => {
-  let controller: OauthController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [OauthController],
-      providers: [OauthService],
-    }).compile();
-
-    controller = module.get<OauthController>(OauthController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
