--- src/oauth/oauth.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { OauthService } from './oauth.service';
-
-describe('OauthService', () => {
-  let service: OauthService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [OauthService],
-    }).compile();
-
-    service = module.get<OauthService>(OauthService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
