+++ src/user/dto/sign_in.dto.ts
@@ -1,4 +1,4 @@
 import { PickType } from '@nestjs/swagger';
 import { Users } from '../entities/user.entitiy';
 
-export class SignInDto extends PickType(Users, ['password', 'email']) {}
+export class SignInDto extends PickType(Users, ['email']) {}
