+++ src/user/dto/signup.dto.ts
@@ -4,6 +4,5 @@ import { Users } from '../entities/user.entitiy';
 export class SignUpDto extends PickType(Users, [
   'name',
   'email',
-  'password',
   'nickname',
 ]) {}
